+++ src/main/kotlin/klib/files/Directory.kt
@@ -1,0 +1,4 @@
+package klib.files
+
+class Directory {
+}