+++ src/test/kotlin/devtests/MutableList.kt
@@ -1,0 +1,9 @@
+package devtests
+
+import klib.extensions.binSearch
+
+fun main() {
+    val list = mutableListOf(1,4,7,9,11,66,99)
+
+    println(list.binSearch(7))
+}